Update logic to handle validating invalid classes
